@@ -1,10 +1,10 @@
 name: mutex
 description: Mutual exclusion with implementation of normal and read-write mutex
-version: 1.1.0
+version: 2.0.0-nullsafety.0
 homepage: https://github.com/hoylen/dart-mutex
 
 environment:
-  sdk: '>=1.0.0 <3.0.0'
+  sdk: '>=2.12.0-0 <3.0.0'
 
 #dependencies:
 #  lib_name: any
